% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{install_scripts}
\alias{install_scripts}
\title{Install rrqueue scripts}
\usage{
install_scripts(dest, overwrite = TRUE)
}
\arguments{
\item{dest}{Directory to install scripts to, preferably in the \code{$PATH}}

\item{overwrite}{Overwrite any existing file?}
}
\description{
Install rrqueue scripts.  This currently includes a script for
starting a worker process.  The script is extremely simple and
does not need upgrading when rrqueue is upgraded.
}

