% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/observer.R
\name{observer}
\alias{observer}
\title{Creates an observer for an rrqueue}
\usage{
observer(queue_name, redis_host = "127.0.0.1", redis_port = 6379)
}
\arguments{
\item{queue_name}{Name of the queue}

\item{redis_host}{Redis hostname}

\item{redis_port}{Redis port number}
}
\description{
Creates an observer for an rrqueue
}

