% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/controller.R
\name{queue}
\alias{queue}
\title{Create an rrqueue queue}
\usage{
queue(name, packages = NULL, sources = NULL, con = NULL)
}
\arguments{
\item{name}{Queue name}

\item{packages}{Character vector of packages to load}

\item{sources}{Character vector of files to source}

\item{con}{Connection to a redis database}
}
\description{
Create an rrqueue queue
}

