% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/rrqlapply.R
\name{rrqlapply}
\alias{rrqlapply}
\alias{rrqlapply_results}
\alias{rrqlapply_submit}
\title{Parallel version of lapply using Redis queuing}
\usage{
rrqlapply(X, FUN, rrq, period = 1, delete_tasks = FALSE)

rrqlapply_submit(X, FUN, rrq)

rrqlapply_results(obj, period = 1, delete_tasks = FALSE)
}
\arguments{
\item{X}{A vector}

\item{FUN}{The name of a function to apply to each element of the
list.  \emph{this will change!}.}

\item{rrq}{An rrq object}

\item{period}{Period to poll for completed tasks.  Affects how
responsive the function is to quiting, mostly.}

\item{delete_tasks}{Delete tasks on successful finish?}
}
\description{
Parallel version of lapply using Redis queuing
}

