% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/task_bundle.R
\name{task_bundle}
\alias{task_bundle}
\title{Create a task bundle}
\usage{
task_bundle(obj, tasks, groups = NULL, names = NULL)
}
\arguments{
\item{obj}{An observer or queue object}

\item{tasks}{A list of tasks}

\item{groups}{Optional vector of groups, may be dropped soon}

\item{names}{Optional vector of names to label output with.}
}
\description{
Create a task bundle.  Generally these are not created manually,
but this page serves to document methods that bundles have.
}

