% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/worker_group.R
\name{worker_group}
\alias{worker_group}
\title{Create group of rrqueue workers}
\usage{
worker_group(queue_name, n, logfile_fmt, poll = 10, con = NULL)
}
\arguments{
\item{queue_name}{Queue name}

\item{n}{Number of processes to start}

\item{logfile_fmt}{Format for the log file, containing a
percentage-d for \code{sprintf}}

\item{poll}{Frequency, in seconds, to poll for updates}

\item{con}{Redis connection.  If this is not NULL probably nothing
good will happen.}
}
\description{
Create group of rrqueue workers
}

