% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/worker_spawn.R
\name{worker_spawn}
\alias{worker_spawn}
\title{Spawn a worker}
\usage{
worker_spawn(queue_name, logfile, redis_host = "127.0.0.1",
  redis_port = 6379, timeout = 20, time_poll = 3,
  heartbeat_period = NULL, heartbeat_expire = NULL)
}
\arguments{
\item{queue_name}{Name of the queue to connect to}

\item{logfile}{Name of a log file to write to (consider \code{tempfile()})}

\item{redis_host}{Host name/IP for the Redis server}

\item{redis_port}{Port for the Redis server}

\item{timeout}{Time to wait for the worker to appear}

\item{time_poll}{Period to poll for the worker}

\item{heartbeat_period}{Period between heartbeat pulses}

\item{heartbeat_expire}{Time that heartbeat pulses will persist}
}
\description{
Spawn a worker in the background
}

