% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/env.R
\name{yaml_env}
\alias{yaml_env}
\title{Load environment variables from a yaml file}
\usage{
yaml_env(filename, section = NULL)
}
\arguments{
\item{filename}{Name of the file}

\item{section}{An optional section of the file to load}
}
\description{
Load environment variables from a yaml file.  This is a hack for a
project.  It may change and may move package.  \code{callr} would
be a better fit probably, but \code{callr} doesn't pull in
\code{yaml} yet so I don't know that it's a good fit.
}
\details{
The yaml file must be sets of key/value pairs of simple data
types.  Something like:

\preformatted{
REDIS_HOST: localhost
}

Alternatively, for use with section, add an extra layer of nesting:

\preformatted{
local:
  REDIS_HOST: localhost
remote:
  REDIS_HOST: redis.marathon.mesos
}
}
