% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/controller.R
\name{queue}
\alias{queue}
\title{Create an rrqueue queue}
\usage{
queue(queue_name, packages = NULL, sources = NULL,
  redis_host = "127.0.0.1", redis_port = 6379, clean = FALSE)
}
\arguments{
\item{queue_name}{Queue name}

\item{packages}{Character vector of packages to load}

\item{sources}{Character vector of files to source}

\item{redis_host}{Redis hostname}

\item{redis_port}{Redis port number}

\item{clean}{Delete any rements of existing queues on startup
(this can cause things to go haywire if processes are still live
working on jobs as they'll clobber your queue).}
}
\description{
Create an rrqueue queue
}

