% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rrqlapply.R
\name{rrqlapply}
\alias{rrqlapply}
\alias{rrqlapply_submit}
\title{Parallel version of lapply using Redis queuing}
\usage{
rrqlapply(X, FUN, rrq, ..., group = NULL, timeout = Inf,
  time_poll = 1, delete_tasks = FALSE, progress_bar = TRUE,
  env = parent.frame())

rrqlapply_submit(X, FUN, rrq, ..., group = NULL, progress_bar = TRUE,
  env = parent.frame())
}
\arguments{
\item{X}{A vector}

\item{FUN}{The name of a function to apply to each element of the
list.  \emph{this will change!}.}

\item{rrq}{An rrq object}

\item{...}{Additional arguments passed to \code{FUN}}

\item{group}{Name of a group for generated task ids.  If not
included, an ID will be generated.}

\item{timeout}{Total length of time to wait for tasks to be
completed.  The default is to wait forever (like \code{lapply}).}

\item{time_poll}{Time to poll for tasks.  Must be an integer.
Because of how the function is implemented, R will be
unresponsive for this long each iteration (unless results are
returned), so the default of 1s should be reasonable.}

\item{delete_tasks}{Delete tasks on successful finish?}

\item{progress_bar}{Display a progress bar?}

\item{env}{Environment to look in.}
}
\description{
Parallel version of lapply using Redis queuing
}
