% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/worker.R
\name{worker}
\alias{worker}
\title{Create an rrqueue worker}
\usage{
worker(queue_name, con = NULL, task_timeout = 60, heartbeat_period = 10,
  heartbeat_expire = 30)
}
\arguments{
\item{queue_name}{Queue name}

\item{con}{Connection to a redis database}

\item{task_timeout}{Task_Timeout for the blocking connection}
}
\description{
Create an rrqueue worker.  This blocks the main loop.
}

