% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/worker.R
\name{worker}
\alias{worker}
\title{Create an rrqueue worker}
\usage{
worker(queue_name, con = NULL, timeout = 60)
}
\arguments{
\item{queue_name}{Queue name}

\item{con}{Connection to a redis database}

\item{timeout}{Timeout for the blocking connection}
}
\description{
Create an rrqueue worker.  This blocks the main loop.
}

